+++ frontend/src/components/index.ts
@@ -7,3 +7,4 @@ export { default as CommonHeader } from './CommonHeader';
 export { default as ServiceMenu } from './ServiceMenu';
 export { default as InputLabel } from './InputLabel';
 export { default as CardSlider } from './CardSlider';
+export { default as TabBar } from './TabBar';
