+++ frontend/src/pages/Detail/index.tsx
@@ -1,9 +1,8 @@
-import { useState } from 'react';
 import styled from 'styled-components';
 import Picture from '../../asset/picture2.png';
 import { ArrowRight } from '../../asset';
 import COLOR from '../../constants/color';
-import { SaleSchedule } from '../../components';
+import { SaleSchedule, TabBar } from '../../components';
 import CommonHeader from '../../components/CommonHeader';
 import { useNavigate } from 'react-router-dom';
 
@@ -16,13 +15,6 @@ const StyledTitleWrapper = styled.div`
   display: flex;
   flex-direction: column;
   margin: 1rem;
-`;
-
-const StyledContentWrapper = styled.div`
-  display: flex;
-  justify-content: space-evenly;
-  padding-top: 1.5rem;
-  border-top: 4px solid ${COLOR.LIGHT_010};
 `;
 
 const StyledLocationWrapper = styled.div`
@@ -47,28 +39,7 @@ const StyledDate = styled.div`
 
 const StyledImg = styled.img``;
 
-export const StyledMenuTitle = styled.li<{ check?: string }>`
-  width: 25%;
-  font-size: 0.9rem;
-  padding-bottom: 1rem;
-  justify-content: center;
-  text-align: center;
-  font-family: 'Pretendard-Regular';
-  display: flex;
-  color: lightgray;
-  &:hover {
-    cursor: pointer;
-  }
-  &#${({ check }) => check} {
-    font-weight: bold;
-    color: ${COLOR.PRI_MAIN};
-    border-bottom: 2px solid ${COLOR.PRI_MAIN};
-  }
-`;
-
 const Detail = () => {
-  const [menuCheck, setMenuCheck] = useState<number>(0);
-  const [check, setCheck] = useState('schedule');
   const menu = [
     { name: '분양일정', option: 'schedule' },
     { name: '공급대상', option: 'target' },
@@ -81,21 +52,7 @@ const Detail = () => {
     2: <>임대조건 데이터 연동중이에요!</>,
     3: <SaleSchedule />,
   };
-  const menuList = menu.map((i, index) => (
-    <StyledMenuTitle
-      key={i.option}
-      id={i.option}
-      check={check}
-      onClick={(e) => clickFunc(e, index)}
-    >
-      {i.name}
-    </StyledMenuTitle>
-  ));
 
-  const clickFunc = (e: any, index: number) => {
-    setMenuCheck(index);
-    setCheck(e.target.id);
-  };
   let navigate = useNavigate();
   return (
     <>
@@ -110,8 +67,7 @@ const Detail = () => {
           <StyledDate>2022.05.15 등록</StyledDate>
         </StyledTitleWrapper>
 
-        <StyledContentWrapper>{menuList}</StyledContentWrapper>
-        {checkList[menuCheck]}
+        <TabBar menu={menu} checkList={checkList} />
       </StyledWrapper>
     </>
   );
